Fix imports
Fix modules and relative imports logic for installation
