--- SlicerPKDIA/SlicerPKDIALib/__init__.py
@@ -1,4 +1,0 @@
-from .InstallLogic import InstallLogic
-from .SegmentationLogic import SegmentationLogic
-from .Signal import Signal
-from .Widget import Widget
