--- SlicerPKDIA/SlicerPKDIALib/pkdia/datasets/__init__.py
@@ -1,1 +1,0 @@
-from .dataset_genkyst import tiny_dataset_genkyst_prod
