--- SlicerPKDIA/SlicerPKDIALib/pkdia/exams/__init__.py
@@ -1,1 +1,0 @@
-from .exam_genkyst_prod import exam_genkyst_prod
