--- SlicerPKDIA/SlicerPKDIALib/pkdia/manage/__init__.py
@@ -1,1 +1,0 @@
-from .manage_genkyst import extract_genkyst_slice_prod
