--- SlicerPKDIA/SlicerPKDIALib/pkdia/nets/__init__.py
@@ -1,1 +1,0 @@
-from .swinv2Unet import SwinV2TwoDecoder
