--- SlicerPKDIA/SlicerPKDIALib/pkdia/utils/__init__.py
@@ -1,7 +1,0 @@
-from .utils import (
-    ModalityEnum,
-    get_array_affine_header,
-    getLargestConnectedArea,
-    normalization_imgs,
-    prob2mask,
-)
