+++ SlicerPKDIA/SlicerPKDIALib/pkdia/utils/utils.py
@@ -1,24 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from enum import Enum, EnumMeta
-
 import numpy as np
 from skimage.measure import label
-
-
-class ModalityEnumMeta(EnumMeta):
-    def __contains__(cls, item):
-        try:
-            cls(item)
-        except ValueError:
-            return False
-        return True
-
-
-class ModalityEnum(str, Enum, metaclass=ModalityEnumMeta):
-    T2 = "MRI T2"
-    CT = "CT"
 
 
 def normalization_imgs(imgs):
